Implement the rest of the site.
I finished this project by implementing the rest of the functionality as well as finishing the CSS styling.
The site has a darker theme to suggest the vastness and emptiness of space.
I just a script to check the password fields to see if they match. If they do, then the error class is toggled and the error message is hidden.